webrender: refactor create frame logic
